@@ -3,30 +3,47 @@
   "version": "1.0.3",
   "description": "A tiny cross-platform promise based wrapper around child_process.spawn.",
   "license": "MIT",
+  "types": "dist/index.d.ts",
+  "main": "dist/index.js",
+  "module": "dist/index.mjs",
+  "files": [
+    "dist"
+  ],
   "repository": {
     "url": "https://github.com/tiaanduplessis/shell-exec",
     "type": "git"
   },
-  "homepage": "https://github.com/tiaanduplessis/shell-exec",
-  "bugs": "https://github.com/tiaanduplessis/shell-exec",
-  "main": "dist/index.js",
-  "files": [
-    "dist/index.js"
-  ],
-  "scripts": {
-    "test": "jest",
-    "lint": "standard --fix dist/index.js",
-    "start": "yarn test",
-    "pretest": "tsc && yarn lint",
-    "precommit": "yarn test"
+  "homepage": "https://github.com/tiaanduplessis/shell-exec#readme",
+  "bugs": {
+    "url": "https://github.com/tiaanduplessis/shell-exec/issues"
   },
+  "engines": {
+    "node": ">=12"
+  },
+  "scripts": {
+    "test": "vitest",
+    "coverage": "vitest run --coverage",
+    "build": "rm -rf dist && tsup src/index.ts --format cjs,esm --dts --target node12",
+    "prepublishOnly": "pnpm run build",
+    "types:check": "tsc --noEmit",
+    "format": "pretty-quick --ignore-path .gitignore",
+    "start": "tsx example.ts"
+  },
+  "keywords": [
+    "shell-exec",
+    "shell",
+    "sh",
+    "exec"
+  ],
   "author": "Tiaan du Plessis",
   "devDependencies": {
     "@types/jest": "^27.4.0",
     "@types/node": "^17.0.15",
-    "husky": "3.0.5",
-    "jest": "24.9.0",
-    "standard": "12.0.1",
-    "typescript": "^4.5.5"
+    "prettier": "^2.6.2",
+    "pretty-quick": "^3.1.3",
+    "tsup": "^5.12.8",
+    "tsx": "^3.1.0",
+    "typescript": "^4.6.4",
+    "vitest": "^0.12.6"
   }
 }
